Attempt at restoring inventory access after TPs/crossings. RemoveClient in Scene was being too aggressive at nixing the user out of the cache. We're now relying on NeedSceneCacheClear to decide whether to nix it or not. All other mods in other files are for better debugging messages.

+++ OpenSim/Region/Environment/Scenes/Scene.cs
@@ -2709,7 +2709,11 @@ namespace OpenSim.Region.Environment.Scenes
                 m_sceneGraph.removeUserCount(!childagentYN);
                 RemoveCapsHandler(agentID);
 
-                CommsManager.UserProfileCacheService.RemoveUser(agentID);
+                if (avatar.Scene.NeedSceneCacheClear(avatar.UUID))
+                {
+                    m_log.InfoFormat("[SCENE]: User {0} is going to another region, profile cache removed in {1}", avatar.UUID, RegionInfo.RegionName);
+                    CommsManager.UserProfileCacheService.RemoveUser(agentID);
+                }
 
                 if (!avatar.IsChildAgent)
                 {
